Add files via upload
The Perceptron is amazing because of the following reasons:

(a) It was developed in 1957, when no one knew or cared about ML.
(b) It was not just an algorithm, people even built a machine to implement it! 
(c) It was one of the first algorithms which learnt from it’s mistakes. No one called it ML then, but it was.
(d) It is just a 9 line algorithm, but each of those nine lines is quite profound.

Perceptrons provide an intuitive understanding of how machines learn from data. In that sense, they are the stepping stones to ML as we know today: neural networks, transformers etc.